feat: put and get published collab data

+++ src/biz/workspace/ops.rs
@@ -18,8 +18,9 @@ use database::resource_usage::get_all_workspace_blob_metadata;
 use database::user::select_uid_from_email;
 use database::workspace::{
   change_workspace_icon, delete_from_workspace, delete_published_collab, delete_workspace_members,
-  get_invitation_by_id, insert_or_replace_publish_collab_meta, insert_user_workspace,
-  insert_workspace_invitation, rename_workspace, select_all_user_workspaces,
+  get_invitation_by_id, insert_or_replace_publish_collab_meta,
+  insert_or_replace_published_collab_blob, insert_user_workspace, insert_workspace_invitation,
+  rename_workspace, select_all_user_workspaces, select_published_collab_blob,
   select_user_is_collab_publisher, select_user_is_workspace_owner, select_workspace,
   select_workspace_invitations_for_user, select_workspace_member, select_workspace_member_list,
   select_workspace_settings, select_workspace_total_collab_bytes, update_updated_at_of_workspace,
@@ -139,32 +140,34 @@ pub async fn publish_workspace_collab(
   Ok(())
 }
 
+pub async fn put_published_collab_blob(
+  pg_pool: &PgPool,
+  workspace_id: &Uuid,
+  doc_name: &str,
+  user_uuid: &Uuid,
+  collab_data: &[u8],
+) -> Result<(), AppError> {
+  check_workspace_owner_or_publisher(pg_pool, user_uuid, workspace_id, doc_name).await?;
+  insert_or_replace_published_collab_blob(pg_pool, workspace_id, doc_name, collab_data).await?;
+  Ok(())
+}
+
+pub async fn get_published_collab_blob(
+  pg_pool: &PgPool,
+  workspace_id: &Uuid,
+  doc_name: &str,
+) -> Result<Vec<u8>, AppError> {
+  select_published_collab_blob(pg_pool, workspace_id, doc_name).await
+}
+
 pub async fn delete_published_workspace_collab(
   pg_pool: &PgPool,
   workspace_id: &Uuid,
   doc_name: &str,
   user_uuid: &Uuid,
 ) -> Result<(), AppError> {
-  let is_owner = select_user_is_workspace_owner(pg_pool, user_uuid, workspace_id).await?;
-  if !is_owner {
-    let is_publisher =
-      select_user_is_collab_publisher(pg_pool, user_uuid, workspace_id, doc_name).await?;
-    if !is_publisher {
-      return Err(AppError::UserUnAuthorized(
-        "User is not the owner of the workspace or the publisher of the document".to_string(),
-      ));
-    }
-  }
-
+  check_workspace_owner_or_publisher(pg_pool, user_uuid, workspace_id, doc_name).await?;
   delete_published_collab(pg_pool, workspace_id, doc_name).await?;
-  Ok(())
-}
-
-pub async fn publish_workspace_collab_data(
-  _pg_pool: &PgPool,
-  _workspace_id: &Uuid,
-  _view_id: &Uuid,
-) -> Result<(), AppError> {
   Ok(())
 }
 
@@ -590,5 +593,24 @@ async fn check_workspace_namespace(new_namespace: &str) -> Result<(), AppError> 
     }
   }
 
+  Ok(())
+}
+
+async fn check_workspace_owner_or_publisher(
+  pg_pool: &PgPool,
+  user_uuid: &Uuid,
+  workspace_id: &Uuid,
+  doc_name: &str,
+) -> Result<(), AppError> {
+  let is_owner = select_user_is_workspace_owner(pg_pool, user_uuid, workspace_id).await?;
+  if !is_owner {
+    let is_publisher =
+      select_user_is_collab_publisher(pg_pool, user_uuid, workspace_id, doc_name).await?;
+    if !is_publisher {
+      return Err(AppError::UserUnAuthorized(
+        "User is not the owner of the workspace or the publisher of the document".to_string(),
+      ));
+    }
+  }
   Ok(())
 }
